// SolitaireLib/Games/AcesUp/Model/IGame.ts
import { IGameBase } from "~CardLib/Model/IGameBase";
import { IPile } from "~CardLib/Model/IPile";
import { GameOptions } from "./GameOptions";

export interface IGame extends IGameBase {
    readonly options: GameOptions;
    readonly stock: IPile;
    readonly foundation: IPile;
    readonly tableaux: IPile[];
}